Add mpd.conf
I should really throw this into beets, or setup something automatic.

I briefly considered putting all my music on 3sirah but that means I
need to constantly be connected to home, which is okay but I don't
generally have need for space locally, so I don't think it's worth the
effort or the lag, so local it stays.

That being said this has been on the backburner for a while. I'm
thinking of just using gmusic or (gasp) spotify but for now I'll stick
with gmpc