Move to common Xmodmap
Commmon in that this is the setup I use on my mac: right alt sends
alt, there's a modifier that also sends control a bit to the right,
easily accessible with a thumb. On an x250, this is prtsc

Additionally, don't bother with parenthesis: the only time I write a
bunch of those is when I'm writing lisp, and though I will likely
write a lot of them I can just get used to the standard, or use
paredit or something.