Have right alt be control
It's pretty convenient to have control be the primary modifier key,
despite alt being in the better keyboard position. So change things
up.

In MacOS, I have the ability to use iTerm2 to remap the right command
key to control. The only thing I'd have to deal with is having to
remember that right command isn't right control everywhere, just in
the terminal, but that's where I spend the majority of my time anyway,
emacs and a web browser :)